+++ internal/cli/command.go
@@ -83,6 +83,7 @@ func Execute(version string, args []string, stdout, stderr io.Writer) error {
 	root.AddCommand(explainCommand(&ran))
 	root.AddCommand(redactCommand(&ran))
 	root.AddCommand(diffCommand(&ran))
+	root.AddCommand(driftCommand(&ran))
 	root.AddCommand(reportCommand(&ran))
 	root.SetArgs(args)
 	selected, err := root.ExecuteC()